There is a problem if you choose to only show upper menus on some pages which is it moves the logo.
Added height to the menus to stop the containers pushing the logo down and added bottom padding to the upper header which does push the header down.

Could add this as a setting, without it things come right up to the dividing line in the header which is not good.