Fix #597 The previous code was adding divs inside the loop used to build multiples messages.
The $msg must be built in the loop before being inserted into the rest of the code.

As line 754 needs to call the setting directly that has to be contained within the loop also.

I think part of the problem with this previously may have been the line that used regex to strip out incorrect tags, previously the list (li) tag was being used and the regex stripped out all other tags including "p". So when the code was switched to require "p" tags this regex should have been updated to strip out li and not p.